+++ config/di.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
+use Psr\EventDispatcher\EventDispatcherInterface;
 use YiiRocks\Voyti\Gdpr\Controller\Privacy\PrivacyController;
+use YiiRocks\Voyti\Gdpr\Service\AnonymizeUserService;
+use YiiRocks\Voyti\Gdpr\Service\GdprExportService;
 use Yiisoft\Translator\CategorySource;
 use Yiisoft\Translator\Message\Php\MessageSource;
 use Yiisoft\Translator\SimpleMessageFormatter;
@@ -10,15 +13,18 @@ use Yiisoft\Translator\SimpleMessageFormatter;
 /** @var array $params */
 
 return [
-    PrivacyController::class => [
-        'class' => PrivacyController::class,
-        '__construct()' => [
-            'gdprConfig' => [
-                'gdprExportProperties' => $params['yiirocks/voyti']['gdpr']['gdprExportProperties'] ?? [],
-                'gdprAnonymizePrefix' => $params['yiirocks/voyti']['gdpr']['gdprAnonymizePrefix'] ?? 'GDPR',
-            ],
-        ],
-    ],
+    PrivacyController::class => PrivacyController::class,
+
+    AnonymizeUserService::class => fn(
+        EventDispatcherInterface $eventDispatcher,
+    ) => new AnonymizeUserService(
+        $eventDispatcher,
+        $params['yiirocks/voyti']['gdpr']['gdprAnonymizePrefix'] ?? 'GDPR',
+    ),
+
+    GdprExportService::class => new GdprExportService(
+        $params['yiirocks/voyti']['gdpr']['gdprExportProperties'] ?? [],
+    ),
 
     // Translation category source for this package's own message files.
     'yiirocks/voyti-gdpr.translator' => [